Fix queries with LEFT JOIN statements
Because we use the respective entities both for table creation and for
joining queries, the table has columns that are not actually populated
on insert/update, and as Room uses the first match, it'll fill in those
empty columns instead of the joined projection values.

We should probably fix this by using a separate class for the joined
data (see [1]), but for now, just spell out the required columns
explicitly.

[1] https://developer.android.com/training/data-storage/room/accessing-data.html#multiple-tables